services/vision_selector/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
torch
transformers
faiss-cpu
numpy
Pillow
